refactor into connection
